Gavin - idk

+++ src/main/view/game/UnitOverviewPanel.java
@@ -155,7 +155,7 @@ public class UnitOverviewPanel extends OverviewPanel{
 				armyId = "No Army";
 			}
 			unitList.add(new UnitItem(unitType, (int)unit.getCurrentHealth(), stats.getOffPow(), stats.getDefPow(),
-					stats.getArmor(), (int)stats.getUpkeep(), armyId));
+					stats.getArmor(), stats.getUpkeep(), armyId));
 			}
 	}
 
@@ -176,13 +176,13 @@ public class UnitOverviewPanel extends OverviewPanel{
 		private final SimpleStringProperty  upkeepProp;
 		private final SimpleStringProperty  armyProp;
 		
-		public UnitItem(String unitType, int health, int attack, int defense, int armor, int upkeep, String army) {
+		public UnitItem(String unitType, int health, int attack, int defense, int armor, double upkeep, String army) {
 			unitTypeProp = new SimpleStringProperty(unitType);
 			healthProp = new SimpleStringProperty(Integer.toString(health));
 			attackProp = new SimpleStringProperty(Integer.toString(attack));
 			defenseProp = new SimpleStringProperty(Integer.toString(defense));
 			armorProp = new SimpleStringProperty(Integer.toString(armor));
-			upkeepProp = new SimpleStringProperty(Integer.toString(upkeep));
+			upkeepProp = new SimpleStringProperty(Double.toString(upkeep));
 			armyProp = new SimpleStringProperty(army);
 		}
